requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Werkzeug==3.0.1
python-socketio==5.10.0
eventlet==0.33.3
openpyxl==3.1.2
python-dotenv>=1.0.0
pytest-playwright>=0.5.0
playwright>=1.44.0

